Drop the changeIP (`reconnect`) service
In ROS2, the gripper has a single IP address at startup.
This service seems irrelevant for these driver's use cases and adds complexity
that's difficult to test and maintain.

We get it back once we really need this.
The driver _does_ reconnect automatically when the connection drops, but
it does it without assuming another IP.