Fix the computation of actual position and speed
We now return the actual data, not the metadata.
Also, we now interpret target position and target speed in the plc
commands as unsigned integers (position in um, speed in um/s) as stated
in the gripper's TCP/IP manual. Only the data in `self.data` need to be stored
and returned as floats.

Check this manual for further details:
[1]: https://stb.cloud.schunk.com/media/IM0046706.PDF