Prevent race conditions between publishing and deletion
This is hard to test deterministically because it's difficult to make
the GIL continue with publishing just after the publishers have been
destroyed. Without locks, it does happen, but only after minutes of
running the test, which is not practical for the local
testing/developing workflow.

Instead, this adds a test that covers deadlocks.