Use correct `device_id` when starting cyclic updates
We erroneously started the cyclic updates before the grippers were connected.
This lead to a Modbus _broadcast_ with the gripper's default constructed
`device_id = 0`.
Unfortunately, this _did_ work for single-gripper setups where only one
gripper would answer, and the error was not noticed.

We now start the cyclic update _after_ connecting the gripper
and thus make sure that the correct `device_id` is used.